pyproject.toml
[project]
name = "lattice-base"
version = "0.1.0"
description = "Core lattice model + CLI for project/task dependency graphs."
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0.0",
  "pyyaml>=6.0.0",
  "pytest>=8.0"
]

[project.scripts]
lattice-base-init = "lattice_base.cli:main_init"
lattice-base-next = "lattice_base.cli:main_next"
lattice-base-validate = "lattice_base.cli:main_validate"
lattice-base-mermaid = "lattice_base.cli:main_mermaid"
lattice-base-test = "lattice_base.cli:main_test"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
